backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
groq